refactor: routes profile 모듈화
Related to: #143
