+++ server/src/routes/api/index.ts
@@ -1,11 +1,4 @@
-import path from 'path';
-import dotenv from 'dotenv';
-dotenv.config({
-  path: path.resolve(__dirname, '../../config/.env.development')
-});
-import express, { Request, Response, NextFunction } from 'express';
-import dbManager from '../../service/dbManager';
-import { DBUser, IProfile } from '../../types/interface';
+import express from 'express';
 
 import isLogin from './isLogin';
 import users from './users';
@@ -15,6 +8,7 @@ import image from './image';
 import comments from './comments';
 import problems from './problems';
 import groups from './groups';
+import profile from './profile';
 
 const router = express.Router();
 
@@ -49,29 +43,7 @@ router.get('/groups/joined/:useridx', groups.joined);
 router.get('/groups/usernum/:groupidx', groups.userNum);
 router.post('/joingroup/:useridx/:postidx', groups.joinLeave);
 
-router.put(
-  '/profile/:useridx',
-  async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      const userIdx = Number(req.params.useridx);
-      const userUpdateData: IProfile = req.body;
-      await dbManager.updateProfile(userUpdateData, userIdx);
-      res.json({ check: true });
-    } catch (err) {
-      console.error(err);
-      res.json({ check: false });
-    }
-  }
-);
-
-router.get(
-  '/profile/:username',
-  async (req: Request, res: Response, next: NextFunction) => {
-    const name: string = req.params.username;
-    const userdata: DBUser = await dbManager.getProfile(name);
-    if (userdata === undefined) res.json({ data: '', error: true });
-    else res.json({ data: userdata, error: false });
-  }
-);
+router.put('/profile/:useridx', profile.update);
+router.get('/profile/:username', profile.get);
 
 module.exports = router;
