+++ server/src/routes/api/profile.ts
@@ -1,0 +1,26 @@
+import { Request, Response, NextFunction } from 'express';
+
+import dbManager from '../../service/dbManager';
+import { DBUser, IProfile } from '../../types/interface';
+
+const profile = {
+  update: async (req: Request, res: Response, next: NextFunction) => {
+    try {
+      const userIdx = Number(req.params.useridx);
+      const userUpdateData: IProfile = req.body;
+      await dbManager.updateProfile(userUpdateData, userIdx);
+      res.json({ check: true });
+    } catch (err) {
+      console.error(err);
+      res.json({ check: false });
+    }
+  },
+  get: async (req: Request, res: Response, next: NextFunction) => {
+    const name: string = req.params.username;
+    const userdata: DBUser = await dbManager.getProfile(name);
+    if (userdata === undefined) res.json({ data: '', error: true });
+    else res.json({ data: userdata, error: false });
+  }
+};
+
+export default profile;
